devmate: Enable Web UI to Access Backend API.
Allow the Web UI to dynamically construct the backend API URL using its own
accessible address and port. This ensures that the backend can be accessed when
it shares the same host address but runs on a different port from the web UI.

Updated the JavaScript code in `App.js` to dynamically construct the backend
API URL based on the web UI's own host and protocol, assuming that they share
the host but run on different ports.

Updated `docker-compose.yaml` and Dockerfiles to accept backend port as a build
argument.

Changed the Web UI's port to 80 in `.env` for more standardized web access.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>